+++ features/reviews/components/RatingSummary/index.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/nextjs-vite";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { reviewsQueryKeys } from "@/features/shared/queryKeys/reviews";
 import RatingSummary from ".";
-import { reviewsQueryKeys } from "../../queries/queryKeys";
 
 const DEFAULT_STATISTICS = {
 	averageScore: 4.5,
