+++ features/shared/queryKeys/favorites.ts
@@ -1,8 +1,9 @@
-import { FavoriteMeetingsListKeyParams } from "../types";
+import { FavoriteMeetingsListKeyParams } from "@/features/favorites/types";
 
 const FAVORITES_QUERY_BASE_KEY = ["favorites"] as const;
 
-export const queryKeys = {
+/** 찜하기/찜해제 관련된 경우 = favoritesQueryKey.all */
+export const favoritesQueryKeys = {
 	favorites: {
 		all: FAVORITES_QUERY_BASE_KEY,
 		list: (params: FavoriteMeetingsListKeyParams) =>
